moved the boards-related logic to the store
